api/requirements.txt
docker==4.2.1
Flask==1.1.2
Flask-Assets==2.0
Flask-Cors==3.0.9
gevent==1.4.0
gunicorn==19.9.0
numpy==1.16.3
pandas==1.1.2
requests==2.21.0
pymongo-3.11.0
dnspython==2.0.0